stm32 ticker: rename hal_tick.h in us_ticker_data.h
